requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
openai==1.3.0
python-dotenv==1.0.0
streamlit==1.22.0
numpy==1.24.3
pycountry==22.3.5
python-dateutil==2.8.2
joblib==1.3.0
